add move function

+++ scenarionet/builder/utils.py
@@ -1,4 +1,3 @@
-import pkg_resources  # for suppress warning
 import copy
 import logging
 import os
@@ -6,8 +5,8 @@ import os.path as osp
 import pickle
 import shutil
 from typing import Callable, List
-import tqdm
 
+import tqdm
 from metadrive.scenario.scenario_description import ScenarioDescription
 
 from scenarionet.common_utils import save_summary_anda_mapping
@@ -119,7 +118,9 @@ def move_dataset(
 ):
     if not os.path.exists(from_path):
         raise FileNotFoundError("Can not find dataset: {}".format(from_path))
-    assert not os.path.samefile(from_path, to_path), "to_directory is the same as from_directory. Abort!"
+    if os.path.exists(to_path):
+        assert exist_ok, "to_directory already exists. Set exists_ok to allow turning it into a dataset"
+        assert not os.path.samefile(from_path, to_path), "to_directory is the same as from_directory. Abort!"
     merge_dataset(
         to_path,
         from_path,
@@ -127,4 +128,7 @@ def move_dataset(
         overwrite=overwrite,
         try_generate_missing_file=True,
     )
-    shutil.rmtree(from_path)
+    files = os.listdir(from_path)
+    if ScenarioDescription.DATASET.MAPPING_FILE in files and ScenarioDescription.DATASET.SUMMARY_FILE in files and len(
+            files) == 2:
+        shutil.rmtree(from_path)
